kill oidc well known middleware and move it to static route
